pyproject.toml
[project]
name = "inventory-saas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.1.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-money>=3.5.4",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-spectacular>=0.29.0",
    "nested-multipart-parser>=1.6.0",
    "psycopg[binary]>=3.3.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]
